# Reuse compression writers