@nmengin 2024-10-07
Hey @michelheusschen,

Thank you for your contribution.

Discussing with the other maintainers, we'd like to bring it as a bug fix in v3.1.
If you agree, could you rebase on the branch `v3.1`?