@michelheusschen 2024-10-07
Sure, that's fine with me. I've updated the target branch to v3.1